Wrong selector starts an infinite loop
First of all, thanks for your work, it save my time :D

Because I custom my form or soemtime the structure may change, the default selectector 

eg```> div``` may fail.

So i add a security to prevent an inifite loop. The user will be prompted in the console, but this is just a proposal.